Fixing nomeclature and versioning.
Rename from SolarMinecraft to SunBlockCore.

Versioning revised. Last published version was 1.0.6.
This is now 0.7.0-1.20.1

Revised docs
