+++ README.md
@@ -1,15 +1,19 @@
 # Solar Powered Minecraft
-`Version 1.0.7`
+Version `0.7.0-1.20.1`
 
-The Solar Minecraft Project is a research project at Concordia University Montreal, under the supervision of Dr. Bart 
-Simon & Dr. Darren Wershler. 
+
+
+SunBlock is a research project at Concordia University Montreal, under the supervision of Dr. Bart 
+Simon & Dr. Darren Wershler. Find more at the project [homepage](https://minecraftbloc.milieux.ca/sunblock). 
 
 The project is exploring alternate minecraft game mechanics and play-styles when integrating 
 Real-Time Solar and Power usage data into the minecraft world. 
 
-This forge mod, running on a server, uses the server's stats to alter the gameplay for all players online. 
+SunBlockCore is a forge mod that, on the server-side loads up real-time solar data into minecraft, altering the game 
+and broadcasting it live to users that recieve it on the client-side and play along the material conditions of the server. 
 
 Please see the [docs folder](./docs/) for more documentation.
+
 ## Getting Started 
 - Clone the repository and open a terminal in the root directory of the project.
 - Run `./gradlew build` to fetch and compile Gradle. This should output `Build Successful`. 
@@ -47,7 +51,7 @@ Please see the [docs folder](./docs/) for more documentation.
 ## The So Far...
 The mod is set up to display server data on a heads-up display (HUD). 
 
-The CPU temperature values are fetched from the command line on the server side (see how this is done [here](https://github.com/MC-Bloc/SB1-DataVisMod/blob/main/src/main/java/solarminecraft/services/DataQueryProcess.java)). 
+The CPU temperature values are fetched from the command line on the server side (see how this is done [here](https://github.com/MC-Bloc/SB1-DataVisMod/blob/main/src/main/java/ca/milieux/sunblock/sunblockcore/services/DataQueryProcess.java)). 
 
 The rest of the values related to the Solar server setup (voltages, powers, and currents from PV panel and battery as well as battery remaining) are 
 all fetched from the JSON file being generated by the python script [serialread.py](https://github.com/Solar-Minecraft-Project/solar_server/blob/main/SerialRead/serialread.py).
