+++ gradle.properties
@@ -42,18 +42,18 @@ mapping_version=2023.09.03-1.20.1
 
 # The unique mod identifier for the mod. Must be lowercase in English locale. Must fit the regex [a-z][a-z0-9_]{1,63}
 # Must match the String constant located in the main mod class annotated with @Mod.
-mod_id=solarminecraft
+mod_id=sunblockcore
 # The human-readable display name for the mod.
-mod_name=Solar Minecraft
+mod_name=SunBlockCore
 # The license of the mod. Review your options at https://choosealicense.com/. All Rights Reserved is the default.
 mod_license=All Rights Reserved
 # The mod version. See https://semver.org/
-mod_version=1.0.7
+mod_version=0.7.0-1.20.1
 # The group ID for the mod. It is only important when publishing as an artifact to a Maven repository.
 # This should match the base package used for the mod sources.
 # See https://maven.apache.org/guides/mini/guide-naming-conventions.html
-mod_group_id=solarminecraft
+mod_group_id=ca.milieux.sunblock
 # The authors of the mod. This is a simple text string that is used for display purposes in the mod list.
-mod_authors=Ella Noyes, Shahrom Ali, Stuart Thiel
+mod_authors=Shahrom Ali, Ella Noyes, Stuart Thiel,
 # The description of the mod. This is a simple multiline text string that is used for display purposes in the mod list.
-mod_description=Solar Powered Minecraft
+mod_description=The core mod for SunBlock - the solar-powered minecraft server.
