+++ src/main/java/ca/milieux/sunblock/sunblockcore/services/setup/ClientSetup.java
@@ -1,0 +1,13 @@
+package ca.milieux.sunblock.sunblockcore.services.setup;
+
+import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
+import ca.milieux.sunblock.sunblockcore.application.client.ClientEventHandler;
+import ca.milieux.sunblock.sunblockcore.domain.SolarServerData;
+
+public class ClientSetup {
+    public static SolarServerData serverData;
+    public static void init(final FMLClientSetupEvent event) {
+        MinecraftForge.EVENT_BUS.register(new ClientEventHandler());
+    }
+}
