--- src/main/java/solarminecraft/application/client/ClientModHandler.java
@@ -1,17 +1,0 @@
-package solarminecraft.application.client;
-
-
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.client.event.RegisterKeyMappingsEvent;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.common.Mod;
-import solarminecraft.SolarMinecraft;
-
-@Mod.EventBusSubscriber(modid = SolarMinecraft.MODID, bus = Mod.EventBusSubscriber.Bus.MOD, value = Dist.CLIENT)
-public class ClientModHandler {
-
-    @SubscribeEvent
-    public void registerKeys(RegisterKeyMappingsEvent event) {
-        event.register(KeyBindings.INSTANCE.showSolarStats);
-    }
-}
