--- src/main/resources/assets/solarminecraft/lang/en_us.json
@@ -1,4 +1,0 @@
-{
-  "key.solarminecraft.show_solar_stats": "Show Solar Stats Key",
-  "item.solarminecraft.solar_sword": "Solar-Powered Sword"
-}
