--- src/main/resources/assets/solarminecraft/models/item/solar_sword.json
@@ -1,6 +1,0 @@
-{
-  "parent": "item/handheld",
-  "textures": {
-    "layer0" : "solarminecraft:item/solar_sword"
-  }
-}
