+++ src/main/resources/assets/sunblockcore/lang/en_us.json
@@ -1,0 +1,4 @@
+{
+  "key.sunblockcore.show_solar_stats": "Show Solar Stats Key",
+  "item.sunblockcore.solar_sword": "Solar-Powered Sword"
+}
