+++ src/main/resources/assets/sunblockcore/models/item/solar_sword.json
@@ -1,0 +1,6 @@
+{
+  "parent": "item/handheld",
+  "textures": {
+    "layer0" : "sunblockcore:item/solar_sword"
+  }
+}
