+++ src/main/resources/assets/sunblockcore/textures/item/solar_sword.png
