+++ src/main/resources/pack.mcmeta
@@ -1,6 +1,6 @@
 {
     "pack": {
-        "description": "Solar Minecraft resources",
+        "description": "SunBlockCore resources",
         "pack_format": 15
     }
 }
